ran in to an issue with the membership function
was simply checking if it could be coerced to false, which is also true of the
zero index. I also was setting values that had been seen before to null. Now
they are set to undefined (perhaps I should del? This is a question about how
garbage collection works, and the consequences there of).